requirements.txt
pytest
selenium
requests
pytest-xdist
pytest-html
allure-pytest
jsonschema
sqlalchemy